fix: Routing base case
Explanation:
If we have a route like this '/hello/world/' splitting by slashes:

```
["", "hello", "world", ""]
```

By default, leading slash is trimmed out, so actually we have:

```
["hello", "world", ""]
```

Now the trailing item (empty string) should be a child of the resource
itself (that allow you to handle resources ending in slash). That is
working now.

The second part is the base case. Since leading slash is being trimmed
out, these two cases should be equivalent:

```
r.Resource("") == r.Resource("/")
```

Now it is not! `r.Resource("")` is the `r` itself whereas `r.Resource("/")`
is a child node with the empty string path.


With this fix behavior is consistent in all cases.